moved redux files to redux folder
