+++ .gitignore
@@ -23,7 +23,7 @@ chrome-user-data
 .vscode
 *.swp
 *.swo
-
+.env
 packages/react-devtools-core/dist
 packages/react-devtools-extensions/chrome/build
 packages/react-devtools-extensions/chrome/*.crx
