+++ package.json
@@ -20,12 +20,14 @@
   "homepage": "https://github.com/atlp-rwanda/c2-barefoot-frontend#readme",
   "dependencies": {
     "@material-ui/core": "^4.11.1",
+    "dotenv": "^8.2.0",
     "html-webpack-plugin": "^4.5.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5"
+    "redux": "^4.0.5",
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
