--- src/store.js
@@ -1,9 +1,0 @@
-import allReducers from './reducers'
-import { createStore } from 'redux'
-
-const store = createStore(
-  allReducers,
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-  )
-
-export default store
